Student module changed, all tests are passing

+++ src/main/java/hantczak/studentDataStorage/student/infrastructure/api/StudentController.java
@@ -39,8 +39,8 @@ public class StudentController {
 
     @GetMapping
     public ResponseEntity<StudentResponse> getAllStudents(@RequestParam(value = "sortType", required = false, defaultValue = "NAME_ASC") String studentSortType,
-                                                          @RequestParam(value = "offset", required = false, defaultValue = "0") long offset,
-                                                          @RequestParam(value = "limit", required = false, defaultValue = "20") long limit) {
+                                                          @RequestParam(value = "offset", required = false, defaultValue = "0") int offset,
+                                                          @RequestParam(value = "limit", required = false, defaultValue = "20") int limit) {
         List<StudentDto> studentDtoList = StudentMapper.studentListToStudentDtoList(studentFacade.getSortedStudents(studentSortType,offset,limit));
         return new ResponseEntity<>(new StudentResponse(studentDtoList), HttpStatus.OK);
     }
